Add markup for the Newsletter Sign-Up Form
Related to the task [[FSSS-37] Newsletter component](https://vtex-dev.atlassian.net/browse/FSSS-37).

Implement a basic and unstyled version of the Newsletter Sign-Up Form.

Note that as we don't have the Pattern Library anymore, nor we're using this new component in the existing pages, it's not being currently used, so I left an example in the comments.